Safari-611.1.10/LayoutTests/js/dfg-negative-array-index-expected.txt

Tests that storing into a negative array index works right.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS array[-1] is 42
PASS successfullyParsed is true

TEST COMPLETE
